pyproject.toml
[project]
name = "mujoco-study"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniel-Salvador",email = "daniovictor@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "mujoco (>=3.3.0,<4.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "tyro (>=0.9.18,<0.10.0)",
    "pillow (>=11.2.1,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
viewer = "python -m mujoco.viewer"
denso = "python -m run_simulation --model-path ./mujoco_study/denso_test.xml"
spot = "python ./run_simulation.py --model-path ./mujoco_study/spot_test.xml"
sandbox = "python -m run_simulation --model-path ./mujoco_study/sandbox.xml"
